[#13] :recycle: isActiveMember() 위치 변경 및 테스트 코드 수정
